+++ tests/ha/control_pod/test_control_pod_soft_failures.py
@@ -80,10 +80,8 @@ class TestControlPodSoftFailure:
         cls.node_worker_list = []
         cls.ha_obj = HAK8s()
         cls.deploy_lc_obj = ProvDeployK8sCortxLib()
-        cls.s3_clean = cls.test_prefix = cls.random_time = None
-        cls.s3acc_name = cls.s3acc_email = cls.bucket_name = cls.object_name = None
-        cls.restore_node = cls.deploy = cls.restore_pod = None
-        cls.repl_num = cls.res_taint = cls.user_list = None
+        cls.s3_clean = cls.test_prefix = cls.restore_node = cls.deploy = None
+        cls.restore_pod = cls.repl_num = cls.res_taint = cls.user_list = cls.test_io_prefix = None
         cls.mgnt_ops = ManagementOPs()
         cls.system_random = secrets.SystemRandom()
         cls.rest_iam_user = RestIamUser()
@@ -114,6 +112,19 @@ class TestControlPodSoftFailure:
             pod_prefix=const.CONTROL_POD_NAME_PREFIX)
         cls.container_name = cls.node_master_list[0].get_all_pods_containers(
             pod_prefix=const.CONTROL_POD_NAME_PREFIX)[cls.ctrl_pods[0]][0]
+        set_type, set_name = cls.node_master_list[0].get_set_type_name(pod_name=cls.ctrl_pods[0])
+        LOGGER.info("Got set_type = %s and set_name = %s.", set_type, set_name)
+        resp = cls.node_master_list[0].get_num_replicas(set_type, set_name)
+        assert_utils.assert_true(resp[0], resp)
+        cls.default_replica = int(resp[1])
+        cls.repl_num = int(len(cls.node_worker_list) / 2 + 1)
+        cls.scale_req = True if cls.default_replica < cls.repl_num else False
+        cls.deploy_name = cls.node_master_list[0].get_deployment_name(
+            pod_prefix=const.CONTROL_POD_NAME_PREFIX)[0]
+        LOGGER.info("Default replicas for %s is %s and can be scaled to: %s",
+                    const.CONTROL_POD_NAME_PREFIX, cls.default_replica, cls.repl_num)
+        cls.num_users = HA_CFG["s3_operation_data"]["200_iam_users"]
+        cls.csm_soft_delay = HA_CFG["common_params"]["15sec_delay"]
 
     def setup_method(self):
         """
@@ -133,14 +144,10 @@ class TestControlPodSoftFailure:
         resp = self.ha_obj.check_cluster_status(self.node_master_list[0])
         assert_utils.assert_true(resp[0], resp[1])
         LOGGER.info("Cluster status is online.")
-        self.s3acc_name = f"ha_s3acc_{self.random_time}"
-        self.s3acc_email = f"{self.s3acc_name}@seagate.com"
-        self.bucket_name = f"ha-mp-bkt-{self.random_time}"
-        self.object_name = f"ha-mp-obj-{self.random_time}"
         if not os.path.exists(self.test_dir_path):
             sysutils.make_dirs(self.test_dir_path)
         self.multipart_obj_path = os.path.join(self.test_dir_path, self.test_file)
-        LOGGER.info("Create IAM user")
+        LOGGER.info("Precondition: Create IAM user before soft-failure.")
         self.users = self.mgnt_ops.create_account_users(nusers=1)
         self.s3_clean = self.users
         self.uids = list(self.users.keys())
@@ -151,8 +158,6 @@ class TestControlPodSoftFailure:
         self.header = self.csm_obj.get_headers(self.csm_user, self.csm_passwd)
         LOGGER.info("Successfully created IAM user")
         self.dtm_obj = DTMRecoveryTestLib(self.access_key, self.secret_key, max_attempts=0)
-        self.repl_num = int(len(self.node_worker_list) / 2 + 1)
-        LOGGER.info("Number of replicas can be scaled for this cluster are: %s", self.repl_num)
         self.pid = self.get_pid(pod_name=self.ctrl_pods[0], container_name=self.container_name,
                                 proc_name=const.CONTROL_POD_SVC_NAME)
         LOGGER.debug("PID of cortx-csm-agent process is: %s", self.pid)
@@ -166,12 +171,15 @@ class TestControlPodSoftFailure:
         if self.s3_clean:
             LOGGER.info("Cleanup: Cleaning created s3 accounts and buckets.")
             resp = self.ha_obj.delete_s3_acc_buckets_objects(self.s3_clean)
-            assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.info("Revert back to default single control pod per cluster if more replicas are "
-                    "created.")
-        if len(self.ctrl_pod_list) > 1:
-            resp = self.node_master_list[0].create_pod_replicas(
-                num_replica=1, deploy=const.CONTROL_POD_NAME_PREFIX)
+            if not resp[0]:
+                LOGGER.error("Cleanup: Failed to perform clean up of created IAM users & buckets")
+            LOGGER.info("Cleanup: Successfully cleaned created IAM users & buckets.")
+        if self.scale_req:
+            LOGGER.info(
+                "Cleanup: Revert back to default replica %s for control pod per cluster.",
+                self.default_replica)
+            resp = self.node_master_list[0].create_pod_replicas(num_replica=self.default_replica,
+                                                                deploy=self.deploy_name)
             assert_utils.assert_true(resp[0], resp[1])
 
         LOGGER.info("Changing control pod process delay to 0")
@@ -215,6 +223,132 @@ class TestControlPodSoftFailure:
                                                      decode=True)
         return resp
 
+    # pylint: disable=too-many-statements
+    # pylint: disable-msg=too-many-locals
+    @pytest.mark.ha
+    @pytest.mark.lc
+    @pytest.mark.tags("TEST-45498")
+    def test_45498(self):
+        """
+        Verify IAM user creation/deletion and IOs on already created IAM user while performing
+        control pod soft-failure in loop one by one for N control pods.
+        """
+        LOGGER.info("STARTED: Verify IAM user creation/deletion & IOs on already created IAM user "
+                    "while performing control pod soft-failure in loop 1 by 1 for N control pods.")
+
+        LOGGER.info("Precondition: Create %s iam users for deletion", self.num_users)
+        del_users = self.mgnt_ops.create_account_users(nusers=self.num_users)
+
+        LOGGER.info("Step 1: Perform IOs before soft-failure.")
+        self.test_prefix = 'test-45498'
+        resp = self.ha_obj.ha_s3_workload_operation(s3userinfo=list(self.s3_clean.values())[0],
+                                                    log_prefix=self.test_prefix, skipcleanup=True,
+                                                    nsamples=2, nclients=2)
+        assert_utils.assert_true(resp[0], resp[1])
+        LOGGER.info("Step 1: Performed IOs before soft-failure.")
+
+        if self.scale_req:
+            LOGGER.info("Scale replicas for control pod to %s", self.repl_num)
+            resp = self.node_master_list[0].create_pod_replicas(num_replica=self.repl_num,
+                                                                deploy=self.deploy_name)
+            assert_utils.assert_true(resp[0], resp[1])
+            LOGGER.info("Total number of control pods in the cluster are: %s", self.repl_num)
+
+        iam_output = Queue()
+        io_output = Queue()
+        event = threading.Event()
+
+        LOGGER.info("Step 2: Starting independent background threads for IOs on already created "
+                    "user and IAM user creation/deletion.")
+        LOGGER.info("Step 2.1: Start IAM user creation/deletion in background during control pod "
+                    "soft-failure.")
+
+        args = {'user_crud': True, 'num_users': self.num_users,
+                'output': iam_output, 'del_users_dict': del_users, 'header': self.header}
+        thr_iam = threading.Thread(target=self.ha_obj.iam_bucket_cruds,
+                                   args=(event, ), kwargs=args)
+        thr_iam.daemon = True  # Daemonize thread
+        self.test_io_prefix = 'test-45498-io'
+        args = {'s3userinfo': list(self.s3_clean.values())[0], 'log_prefix': self.test_io_prefix,
+                'nclients': 1, 'nsamples': 8, 'output': io_output, 'setup_s3bench': False}
+        LOGGER.info("Step 2.2: Start IOs in background during control pod soft-failure.")
+        io_thread = threading.Thread(target=self.ha_obj.event_s3_operation, args=(event,),
+                                     kwargs=args)
+        io_thread.daemon = True  # Daemonize thread
+        io_thread.start()
+        thr_iam.start()
+        LOGGER.info("Step 2: Started independent background threads for IOs on already created "
+                    "user and IAM user creation/deletion.")
+
+        LOGGER.info("Step 3: Perform soft-failure of %s control pods in loop and check cluster "
+                    "has failures", self.repl_num)
+        LOGGER.info("Get the list of control pods")
+        pod_list = self.node_master_list[0].get_all_pods(pod_prefix=const.CONTROL_POD_NAME_PREFIX)
+        event.set()
+        for pod in pod_list:
+            _ = self.node_master_list[0].kill_process_in_container(
+                pod_name=pod, container_name=self.container_name, pid=self.pid, safe_kill=True)
+            LOGGER.info("Step 3.1: Check cluster has failures.")
+            resp = self.ha_obj.check_cluster_status(self.node_master_list[0])
+            assert_utils.assert_false(resp[0], resp)
+            LOGGER.info("Step 3.1: Checked cluster has failures.")
+        LOGGER.info("Step 3: Performed soft-failure of %s control pods in loop and checked cluster "
+                    "has failures", self.repl_num)
+
+        time.sleep(self.csm_soft_delay)
+        event.clear()
+        LOGGER.info("Waiting for background threads to join")
+        thr_iam.join()
+        io_thread.join()
+        LOGGER.info("Step 4: Verifying responses from background processes")
+        LOGGER.info("Step 4.1: Verify background process for IAM user creation")
+        iam_resp = tuple()
+        while len(iam_resp) != 4:
+            iam_resp = iam_output.get(timeout=HA_CFG["common_params"]["60sec_delay"])
+        if not iam_resp:
+            assert_utils.assert_true(False, "Background process failed to do IAM user CRUD "
+                                            "operations")
+        failed = iam_resp[1]
+        exp_failed = iam_resp[0]
+        user_del_failed = iam_resp[2]
+        created_users = iam_resp[3]
+        assert_utils.assert_false(failed, "No IAM user creation/deletion expected to fail before "
+                                          f"or after soft-failure: {failed}")
+        assert_utils.assert_true(exp_failed, "IAM user creation/deletion expected to fail during "
+                                             f"soft-failure: {exp_failed}")
+        if created_users:
+            for i_i in created_users:
+                self.s3_clean.update(i_i)
+        LOGGER.info("Updating dict for clean up with remaining users")
+        for i_i in user_del_failed:
+            self.s3_clean.update({i_i: del_users[i_i]})
+        LOGGER.info("Step 4.1: Verified background process for IAM user creation/deletion")
+
+        LOGGER.info("Step 4.2: Verify background process of IOs")
+        responses = {}
+        while len(responses) != 2:
+            responses = io_output.get(timeout=HA_CFG["common_params"]["60sec_delay"])
+        pass_logs = list(x[1] for x in responses["pass_res"])
+        fail_logs = list(x[1] for x in responses["fail_res"])
+        all_logs = pass_logs + fail_logs
+        resp = self.ha_obj.check_s3bench_log(file_paths=all_logs)
+        assert_utils.assert_false(len(resp[1]), f"Logs which contain failures: {resp[1]}")
+        LOGGER.info("Step 4.2: Verified background process of IOs")
+        LOGGER.info("Step 4: Verified responses from background processes")
+
+        LOGGER.info("Step 5: Create IAM user and perform IOs after soft-failure.")
+        users = self.mgnt_ops.create_account_users(nusers=1)
+        self.test_prefix = 'test-45498-1'
+        self.s3_clean.update(users)
+        resp = self.ha_obj.ha_s3_workload_operation(s3userinfo=list(users.values())[0],
+                                                    log_prefix=self.test_prefix, skipcleanup=True,
+                                                    nclients=2, nsamples=2, setup_s3bench=False)
+        assert_utils.assert_true(resp[0], resp[1])
+        LOGGER.info("Step 5: Created IAM user and performed IOs after soft-failure.")
+
+        LOGGER.info("ENDED: Verify IAM user creation/deletion and IOs on already created IAM user "
+                    "while performing control pod soft-failure in loop 1 by 1 for N control pods.")
+
     # pylint: disable-msg=too-many-statements
     @pytest.mark.ha
     @pytest.mark.lc
@@ -229,11 +363,13 @@ class TestControlPodSoftFailure:
 
         proc_restart_delay = HA_CFG["common_params"]["control_proc_restart_delay"]
         num_users = HA_CFG["s3_operation_data"]["no_csm_users"]
-        LOGGER.info("Scale replicas for control pod to %s", self.repl_num)
-        resp = self.node_master_list[0].create_pod_replicas(num_replica=self.repl_num,
-                                                            deploy=const.CONTROL_POD_NAME_PREFIX)
-        assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.info("Total number of control pods in the cluster are: %s", self.repl_num)
+
+        if self.scale_req:
+            LOGGER.info("Scale replicas for control pod to %s", self.repl_num)
+            resp = self.node_master_list[0].create_pod_replicas(num_replica=self.repl_num,
+                                                                deploy=self.deploy_name)
+            assert_utils.assert_true(resp[0], resp[1])
+            LOGGER.info("Total number of control pods in the cluster are: %s", self.repl_num)
 
         LOGGER.info("Adding %s delay in config files of all control pods", proc_restart_delay)
         self.add_proc_restart_delay(proc_restart_delay)
@@ -345,11 +481,12 @@ class TestControlPodSoftFailure:
         num_users = HA_CFG["s3_operation_data"]["iam_users"]
         num_bkts = HA_CFG["s3_operation_data"]["no_bkt_del_ctrl_pod"]
         proc_restart_delay = HA_CFG["common_params"]["min_control_proc_restart_delay"]
-        LOGGER.info("Scale replicas for control pod to %s", self.repl_num)
-        resp = self.node_master_list[0].create_pod_replicas(num_replica=self.repl_num,
-                                                            deploy=const.CONTROL_POD_NAME_PREFIX)
-        assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.info("Total number of control pods in the cluster are: %s", self.repl_num)
+        if self.scale_req:
+            LOGGER.info("Scale replicas for control pod to %s", self.repl_num)
+            resp = self.node_master_list[0].create_pod_replicas(num_replica=self.repl_num,
+                                                                deploy=self.deploy_name)
+            assert_utils.assert_true(resp[0], resp[1])
+            LOGGER.info("Total number of control pods in the cluster are: %s", self.repl_num)
 
         LOGGER.info("Adding %s delay in config files of all control pods", proc_restart_delay)
         self.add_proc_restart_delay(proc_restart_delay)
